kvserver: put snapshot settings into sep file
Epic: CRDB-46488
Release note: none
